@@ -1,37 +1,87 @@
 package com.example.scanshield_mobile_app;
 
+import android.Manifest;
 import android.content.Intent;
-import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-
+import android.widget.Toast;
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
 public class Loading2 extends AppCompatActivity {
 
-    private static final String PREFS_NAME = "UserPrefs";
-    private static final String KEY_ONBOARDING_COMPLETED = "isOnboardingCompleted";
+    private static final int PERMISSION_REQUEST_CODE = 100;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_loading2);
 
-        // Check if onboarding is completed
-        SharedPreferences prefs = getSharedPreferences(PREFS_NAME, MODE_PRIVATE);
-        boolean isOnboardingCompleted = prefs.getBoolean(KEY_ONBOARDING_COMPLETED, false);
+        // Check and request permissions
+        checkAndRequestPermissions();
+    }
 
-        // Transition after 2 seconds
-        new Handler(Looper.getMainLooper()).postDelayed(() -> {
-            Intent intent;
-            if (isOnboardingCompleted) {
-                intent = new Intent(Loading2.this, Home.class);
-            } else {
-                intent = new Intent(Loading2.this, GetStarted1.class);
+    private void checkAndRequestPermissions() {
+        // List of permissions to request
+        String[] permissions = new String[]{
+                Manifest.permission.READ_PHONE_STATE,
+                Manifest.permission.ANSWER_PHONE_CALLS, // Added
+                Manifest.permission.READ_SMS,           // Added for SMS features
+                Manifest.permission.RECEIVE_SMS,       // Added for SMS features
+                Manifest.permission.READ_CALL_LOG,     // Added for call log features
+                Manifest.permission.READ_CONTACTS      // Added for contact features
+        };
+
+        // Check which permissions are not granted
+        boolean allPermissionsGranted = true;
+        for (String permission : permissions) {
+            if (ContextCompat.checkSelfPermission(this, permission) != PackageManager.PERMISSION_GRANTED) {
+                allPermissionsGranted = false;
+                break;
             }
+        }
+
+        if (allPermissionsGranted) {
+            // All permissions are granted, proceed to the next activity
+            proceedToNextActivity();
+        } else {
+            // Request permissions
+            ActivityCompat.requestPermissions(this, permissions, PERMISSION_REQUEST_CODE);
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == PERMISSION_REQUEST_CODE) {
+            boolean allPermissionsGranted = true;
+            for (int result : grantResults) {
+                if (result != PackageManager.PERMISSION_GRANTED) {
+                    allPermissionsGranted = false;
+                    break;
+                }
+            }
+
+            if (allPermissionsGranted) {
+                // All permissions granted, proceed to the next activity
+                proceedToNextActivity();
+            } else {
+                // Some permissions denied, show a message and either proceed or exit
+                Toast.makeText(this, "Required permissions not granted. Some features may not work.", Toast.LENGTH_LONG).show();
+                proceedToNextActivity();
+            }
+        }
+    }
+
+    private void proceedToNextActivity() {
+        new Handler(Looper.getMainLooper()).postDelayed(() -> {
+            Intent intent = new Intent(Loading2.this, Home.class);
             startActivity(intent);
-            finish(); // Close Loading2 activity
-        }, 2000); // 2000ms = 2 seconds
+            finish();
+        }, 2000);
     }
 }
