@@ -1,30 +1,93 @@
 package com.example.scanshield_mobile_app;
 
+import android.Manifest;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.os.Handler;
-
-import androidx.activity.EdgeToEdge;
+import android.telecom.TelecomManager;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import android.util.Log;
 
 public class MainActivity extends AppCompatActivity {
+    private static final String TAG = "Loading2";
+    private static final int PERMISSION_REQUEST_CODE = 1;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_loading2);
 
-        //Loading delay
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent LogoIntent = new Intent(getApplicationContext(), Loading2.class);
-                startActivity(LogoIntent);
+        // Request permissions
+        String[] permissions = {
+                Manifest.permission.READ_PHONE_STATE,
+                Manifest.permission.READ_CALL_LOG,
+                Manifest.permission.READ_CONTACTS,
+                Manifest.permission.ANSWER_PHONE_CALLS,
+                Manifest.permission.CALL_PHONE,
+                Manifest.permission.READ_SMS,
+                Manifest.permission.RECEIVE_SMS
+        };
+
+        if (!hasPermissions(permissions)) {
+            ActivityCompat.requestPermissions(this, permissions, PERMISSION_REQUEST_CODE);
+        } else {
+            setupDefaultDialer();
+        }
+    }
+
+    private void setupDefaultDialer() {
+        // Prompt for default phone app
+        TelecomManager telecomManager = (TelecomManager) getSystemService(TELECOM_SERVICE);
+        if (telecomManager != null && !getPackageName().equals(telecomManager.getDefaultDialerPackage())) {
+            Intent intent = new Intent(TelecomManager.ACTION_CHANGE_DEFAULT_DIALER);
+            intent.putExtra(TelecomManager.EXTRA_CHANGE_DEFAULT_DIALER_PACKAGE_NAME, getPackageName());
+            startActivity(intent);
+        }
+
+        // Proceed to Home activity
+        try {
+            Intent intent = new Intent(this, Home.class);
+            startActivity(intent);
+            finish();
+        } catch (Exception e) {
+            Log.e(TAG, "Failed to start Home activity: " + e.getMessage());
+            // Fallback to MainActivity or show error
+            Intent fallbackIntent = new Intent(this, MainActivity.class);
+            startActivity(fallbackIntent);
+            finish();
+        }
+    }
+
+
+    private boolean hasPermissions(String[] permissions) {
+        for (String permission : permissions) {
+            if (ContextCompat.checkSelfPermission(this, permission) != PackageManager.PERMISSION_GRANTED) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == PERMISSION_REQUEST_CODE && grantResults.length > 0) {
+            boolean allGranted = true;
+            for (int result : grantResults) {
+                if (result != PackageManager.PERMISSION_GRANTED) {
+                    allGranted = false;
+                    break;
+                }
+            }
+            if (allGranted) {
+                setupDefaultDialer();
+            } else {
+                Log.e(TAG, "Required permissions not granted");
+                // Optionally show a message to the user
                 finish();
             }
-        }, 2000);
-
+        }
     }
 }
-
